feat: moves some types into types folder
