+++ src/lib/types/api.ts
@@ -1,0 +1,4 @@
+export type HandleRequestResponse = {
+	psychological?: string | null;
+	legal?: string | null;
+};
